Start to abstract the questions.
