+++ public_html/wp-content/mu-plugins/camptix-tweaks/addons/extra-fields.php
@@ -1,0 +1,23 @@
+<?php
+namespace WordCamp\CampTix_Tweaks;
+use CampTix_Addon;
+
+defined( 'WPINC' ) || die();
+
+/**
+ * Abstraction class for extra fields.
+ */
+abstract class Extra_Field extends CampTix_Addon {
+	const SLUG = '';
+
+	public $label      = '';
+	public $a11y_label = null;
+	public $question   = '';
+	public $options    = array();
+
+	/**
+	 * Hook into WordPress and Camptix.
+	 */
+	public function camptix_init() {
+	}
+}
